@@ -1,3 +1,3 @@
 module go.k6.io/croconf
 
-go 1.16
+go 1.20
